requirements.txt
Flask==3.0.2
load_dotenv